+++ NEWS.md
@@ -3,6 +3,9 @@
 * `use_air()` is a new function to configure a project to use
   [Air](https://posit-dev.github.io/air), an extremely fast R code formatter.
 
+* `use_tidy_style()` is deprecated, in favor of using Air for any new, explicit
+  formatting efforts (#2110).
+
 # usethis 3.1.0
 
 * `use_vignette()` and `use_article()` support Quarto. The `name` of the new